backend/pyproject.toml
[tool.poetry]
name = "familytree-backend"
version = "0.1.0"
description = "Django backend for Family Tree web application"
authors = ["Benny Christanto <niotech5@gmail.com>"]
readme = "README.md"
packages = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
djangorestframework = "^3.15.0"
django-cors-headers = "^4.3.0"
Pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-django = "^4.7.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
